requirements.txt
discord.py==2.4.0
aiohttp==3.9.5
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
